+++ nhd2-exp/src/gui/widget/infobox.cpp
@@ -405,7 +405,7 @@ int CInfoBox::exec(int timeout)
 	if ( timeout == -1 )
 		timeout = g_settings.timing[SNeutrinoSettings::TIMING_EPG];
 
-	unsigned long long timeoutEnd = CRCInput::calcTimeoutEnd( timeout );
+	uint64_t timeoutEnd = CRCInput::calcTimeoutEnd( timeout );
 
 	bool loop = true;
 	while (loop)
